requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
abseil-cpp=20211102.0=he9d5cce_0
altair=4.1.0=py_1
appnope=0.1.2=py310hecd8cb5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310hca72f7f_0
arrow-cpp=8.0.0=py310h72c8010_0
asn1crypto=1.5.1=pyhd8ed1ab_0
attrs=21.4.0=pyhd3eb1b0_0
aws-c-common=0.4.57=hb1e8313_1
aws-c-event-stream=0.1.6=h23ab428_5
aws-checksums=0.1.9=hb1e8313_0
aws-sdk-cpp=1.8.185=he271ece_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py310hecd8cb5_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.4=py310hecd8cb5_0
boost-cpp=1.73.0=hca72f7f_12
bottleneck=1.3.5=py310h4e76f89_0
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
brotlipy=0.7.0=py310hca72f7f_1002
bzip2=1.0.8=h1de35cc_0
c-ares=1.18.1=hca72f7f_0
ca-certificates=2022.07.19=hecd8cb5_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.6.15=py310hecd8cb5_0
cffi=1.15.1=py310hc55c11b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py310hecd8cb5_0
cloudpickle=2.0.0=pyhd3eb1b0_0
commonmark=0.9.1=pyhd3eb1b0_0
cryptography=35.0.0=py310ha82f1d4_2
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py310he9d5cce_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py310hecd8cb5_0
fftw=3.3.9=h9ed2024_1
freetype=2.11.0=hd8bbffd_0
future=0.18.2=py310hecd8cb5_1
gflags=2.2.2=h0a44026_0
giflib=5.2.1=haf1e3a3_0
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.18=pyhd3eb1b0_1
glog=0.5.0=h23ab428_0
gmp=6.2.1=h2e338ed_0
grpc-cpp=1.46.1=h067a048_0
icu=58.2=h0a44026_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py310hecd8cb5_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.15.2=pyh736e0ef_0
ipython=7.31.1=py310hecd8cb5_1
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py310hecd8cb5_1
jinja2=3.0.3=pyhd3eb1b0_0
jpeg=9e=hca72f7f_0
jsonschema=4.4.0=py310hecd8cb5_0
jupyter_client=7.3.4=py310hecd8cb5_0
jupyter_core=4.10.0=py310hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
krb5=1.19.2=hcd88c3b_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libboost=1.73.0=h3fa6bed_12
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libcurl=7.84.0=h6dfd666_0
libcxx=12.0.0=h2f01273_0
libdeflate=1.8=h9ed2024_5
libedit=3.1.20210910=hca72f7f_0
libev=4.33=h9ed2024_1
libevent=2.1.12=h0a4fc7d_0
libffi=3.3=hb1e8313_2
libgfortran=5.0.0=11_2_0_h246ff09_26
libgfortran5=11.2.0=h246ff09_26
libiconv=1.16=hca72f7f_2
libnghttp2=1.46.0=ha29bfda_0
libpng=1.6.37=ha441bb4_0
libprotobuf=3.20.1=h8346a28_0
libsodium=1.0.18=h1de35cc_0
libssh2=1.10.0=h0a4fc7d_0
libthrift=0.15.0=h054ceb0_0
libtiff=4.4.0=h2ef1027_0
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
llvm-openmp=12.0.0=h0dcd299_1
lz4-c=1.9.3=h23ab428_1
markupsafe=2.1.1=py310hca72f7f_0
matplotlib-inline=0.1.6=py310hecd8cb5_0
mistune=0.8.4=py310hca72f7f_1000
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py310hca72f7f_0
mkl_fft=1.3.1=py310hf879493_0
mkl_random=1.2.2=py310hc081a56_0
nbclient=0.5.13=py310hecd8cb5_0
nbconvert=6.4.4=py310hecd8cb5_0
nbformat=5.3.0=py310hecd8cb5_0
ncurses=6.3=hca72f7f_3
nest-asyncio=1.5.5=py310hecd8cb5_0
notebook=6.4.12=py310hecd8cb5_0
numexpr=2.8.3=py310hdcd3fac_0
numpy=1.21.5=py310hdcd3fac_3
numpy-base=1.21.5=py310hfd2de13_3
openssl=1.1.1q=hca72f7f_0
orc=1.7.4=h9274d09_0
oscrypto=1.2.1=pyhd3deb0d_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.3=py310he9d5cce_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
patsy=0.5.2=py310hecd8cb5_1
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py310hde71d04_1
pip=22.1.2=py310hecd8cb5_0
plotly=5.9.0=py310hecd8cb5_0
prometheus_client=0.14.1=py310hecd8cb5_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.20.1=py310he9d5cce_0
psutil=5.9.1=py310h6c45266_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyarrow=8.0.0=py310h53e4f6e_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodomex=3.15.0=py310h6c45266_0
pydeck=0.7.1=py310hecd8cb5_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.4.0=pyhd8ed1ab_0
pympler=0.9=py_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyrsistent=0.18.0=py310hca72f7f_0
pysocks=1.7.1=py310hecd8cb5_0
python=3.10.4=hdfd78df_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.20.0=pyhd8ed1ab_0
python-fastjsonschema=2.16.2=py310hecd8cb5_0
python_abi=3.10=2_cp310
pytz=2022.1=py310hecd8cb5_0
pyzmq=23.2.0=py310he9d5cce_0
re2=2022.04.01=he9d5cce_0
readline=8.1.2=hca72f7f_1
requests=2.28.1=py310hecd8cb5_0
rich=12.5.1=py310hecd8cb5_0
scipy=1.7.3=py310h0a4c7d7_2
semver=2.13.0=pyhd3eb1b0_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=63.4.1=py310hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.1.9=he9d5cce_0
snowflake-connector-python=2.7.11=py310hc081a56_0
soupsieve=2.3.1=pyhd3eb1b0_0
spyder-kernels=2.3.3=py310h2ec42d9_0
sqlite=3.39.2=h707629a_0
statsmodels=0.13.2=py310hca72f7f_0
streamlit=1.11.0=py310hecd8cb5_0
tenacity=8.0.1=py310hecd8cb5_1
terminado=0.13.1=py310hecd8cb5_0
testpath=0.6.0=py310hecd8cb5_0
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.11.2=pyhd3eb1b0_0
tornado=6.1=py310hca72f7f_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.3.0=py310hecd8cb5_0
typing_extensions=4.3.0=py310hecd8cb5_0
tzdata=2022a=hda174b7_0
tzlocal=2.1=py310hecd8cb5_0
urllib3=1.26.11=py310hecd8cb5_0
utf8proc=2.6.1=h9ed2024_0
validators=0.18.2=pyhd3eb1b0_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py310hecd8cb5_1
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py310hecd8cb5_0
wurlitzer=3.0.2=py310hecd8cb5_0
xz=5.2.5=hca72f7f_1
zeromq=4.3.4=h23ab428_0
zipp=3.8.0=py310hecd8cb5_0
zlib=1.2.12=h4dc903c_2
zstd=1.5.2=hcb37349_0
